WIP: Move the database of supported devices out into runtime loaded files
When fwupd is installed in long-term support distros it's very hard to backport
new versions as new hardware is released.

There are several reasons why we can't just include the mapping and quirk
information in the AppStream metadata:

 * The extra data is hugely specific to the installed fwupd plugin versions
 * The device-id is per-device, and the mapping is usually per-plugin
 * Often the information is needed before the FuDevice is created
 * There are security implications in allowing plugins to handle new devices

The idea with FuHwdb is that the end user can drop an additional (or replace
an existing) file in a .d director with a simple format and the hardware will
magically start working. This assumes no new quirks are required, as this would
obviously need code changes, but allows us to get most existing devices working
in an easy way without the user compiling anything.

This allows us to fix issues like https://github.com/hughsie/fwupd/issues/265

FIXMEs:

 * I'm not happy with `FuHwdb` as a name, it's not a DB and too similar to FuHwids
 * Is `.hwdb` a good extension?
 * Could we use a fwupd-installed AppStream file with the quirks in?
 	`org.freedesktop.fwupd.plugin.dfu.quirks.metainfo` and then use `<custom>` and `<value>` --
 	XML isn't super human-write-friendly and kinda verbose...
 * The DFU parts need to be split off as a different commit
 * Other plugins need to be ported, e.g. ebitdo, unifying, purism, and avr32